Added an advanced python version check
Now the python version check don't look for an exact match but on a
compatible version (same or newer version) using a closer match.

On Ubuntu 18.04 the default python3 version is 3.6 and it is not
recommended to change it because it might breaks other programs.
Therefore using  a version check by simply checking the python3
--version was not appropriate because it was rebuilding a python 3.7
even if it was already installed.