Improved and fixed dependencies in makefile
Makefile dependencies where not correct. For example if a pip requirements
file were modified the requirements needed to be manually re-installed.
There was also a bug using the lint target when a local python build
was done. Because the local python build was inside a normal directory,
its content was passed to pylint which then failed. Also the local build
for another python version of 3.7 would have not worked due to hard
coded python binary to python3.7.

Now all dependencies have been corrected, which means that in between
two call of make serve if the requirements.txt has been modified, the
target make setup is automatically called.

The python local build is now done within a hidden directory (.local)
which is then not passed to pylint nor to yapf.

Also make use of proper make target in the buidlspec.