feature_BGDIINF_SB-2056: Removed the route prefix and use SCRIPT_NAME, configure gunicorn for secure scheme header
The path prefix is now set by the deployment using the WSGI SCRIPT_NAME
environment variable. Also configured gunicorn to use secure scheme header which allows to
get rid of the middleware. This settings are needed in order to set the
correct http proto in generated urls in responses.

Changed the default ALLOWED_ORIGIN to allow everything.