Allow any valid @<scale>x and improved route parsing
When using a valid scale not part of 1x, 0.5x or 2x, the service
silently used scale 1x. Also when entering an invalid color value type,
e.g. `@1x-0,0,a.pgn`, the service returned 404 instead of 400.

Now we can use any valid scale and not only 1x, 0.5x and 2x. Also the
parsing of the route don't uses converter anymore which improved the
parsing and allow to use correct error message in case of invalid scale
or invalid color.