BGDIINF_SB-2420: Origin validation fixes - #patch
In the case where the request is done from a subdomain the Origin header and
the Sec-Fetch-Site headers were set. The later was however not set to
'same-origin' but 'same-site' because the origin is not equal but from a subdomain.

Due to this the validation failed because we only allowed Sec-Fetch-Site==same-origin.
This would also have been an issue if we would allow a cross site origin in
the allowed domain, in this case the Sec-Fetch-Site would be 'cross-site'.

So changed the validation logic to be clearer and correct. The validation is
done only on the Origin header with a first fallback to Sec-Fetch-Site if Origin
is not available and a last fallback to Referer header if the previous are
not available.