BGDIINF_SB-2650: Added Env Variables for workers
Issue :
The number of workers couldn't be set by an environment
variable, and the working directory was the default directory, which
leads to disk reading, with poor performances.

Fix :
We created two environment variables, to allow kubernetes to set those
at will, which will allow us to have a tmpfs as a working directory, and
to change the number of workers at will to see if performances improve.

updating README

setting the default value for TMPFS directory